Move CommandSet to PluginFolder
